requirements.txt

# GPU support (optional)
#torch==2.6.0+cu118
#torchaudio==2.5.1+cu118
#torchvision==0.21.0+cu118

# Core dependencies
numpy>=1.20.0
opencv-python>=4.5.4.60
PyYAML>=6.0
ultralytics>=8.3.104  # For YOLOv11
python-dotenv>=1.0.0  # For .env file support
roboflow==1.1.60

# MLflow tracking
mlflow>=2.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Utilities
tqdm>=4.64.0
pillow>=9.0.0
pandas>=1.3.0


